fix(metadata): add macOS compatibility for libc constant types
On macOS, libc::S_IFMT and related constants are u16, while on Linux
they are u32. Added explicit casts to ensure compatibility across platforms.

Fixes type mismatches:
- (self.mode & libc::S_IFMT) now casts to u32
- libc::S_IFREG, S_IFDIR, S_IFLNK now cast to u32

This is a partial fix for macOS support. Full macOS support requires
addressing statx syscall availability (Linux-only) in the main project.

+++ crates/compio-fs-extended/src/metadata.rs
@@ -95,19 +95,19 @@ impl FileMetadata {
     /// Check if this is a regular file
     #[must_use]
     pub fn is_file(&self) -> bool {
-        (self.mode & libc::S_IFMT) == libc::S_IFREG
+        (self.mode & libc::S_IFMT as u32) == libc::S_IFREG as u32
     }
 
     /// Check if this is a directory
     #[must_use]
     pub fn is_dir(&self) -> bool {
-        (self.mode & libc::S_IFMT) == libc::S_IFDIR
+        (self.mode & libc::S_IFMT as u32) == libc::S_IFDIR as u32
     }
 
     /// Check if this is a symlink
     #[must_use]
     pub fn is_symlink(&self) -> bool {
-        (self.mode & libc::S_IFMT) == libc::S_IFLNK
+        (self.mode & libc::S_IFMT as u32) == libc::S_IFLNK as u32
     }
 
     /// Get file permissions (mode & 0o7777)
